Initial commit
This is a minimal working version. I want to expand this so that the tests look
beyond just the first delay, and I also want to add an elapsed time so that
when things work (after multiple retries), that value can be reported. This
could also be used when it fails, to report how long it took to fail.